+++ crates/telio-lana/src/event_log_file.rs
@@ -224,6 +224,25 @@ pub mod moose {
         FileDeleted,
     }
 
+    impl From<TrackerState> for Result {
+        fn from(state: TrackerState) -> Result {
+            match state {
+                TrackerState::Ready => Result::Success,
+                TrackerState::AlreadyStarted => Result::AlreadyInitiated,
+                _ => unreachable!("All other cases are not covered by Result enum"),
+            }
+        }
+    }
+
+    impl From<MooseError> for Error {
+        fn from(moose_error: MooseError) -> Error {
+            match moose_error {
+                MooseError::NotInitiated => Error::NotInitiatedError,
+                _ => Error::EventLogError,
+            }
+        }
+    }
+
     pub trait InitCallback: Send {
         fn after_init(&self, result: &std::result::Result<TrackerState, MooseError>);
     }
@@ -252,7 +271,7 @@ pub mod moose {
         init_cb: Box<(dyn InitCallback + 'static)>,
         error_cb: Box<(dyn ErrorCallback + Sync + std::marker::Send + 'static)>,
     ) -> std::result::Result<Result, Error> {
-        match super::event_log(
+        let result = match super::event_log(
             "init",
             Some(vec![
                 event_path.as_str(),
@@ -262,8 +281,15 @@ pub mod moose {
                 prod.to_string().as_str(),
             ]),
         ) {
-            Ok(_) => Ok(Result::Success),
-            _ => Err(Error::EventLogError),
+            Ok(_) => Ok(TrackerState::Ready),
+            _ => Err(MooseError::NotInitiated),
+        };
+
+        init_cb.after_init(&result);
+
+        match result {
+            Ok(r) => Ok(Result::from(r)),
+            Err(e) => Err(Error::from(e)),
         }
     }